Implement UserDAO
Add tests for UserDAOImpl,
split setup.sql file into setup phases

+++ src/main/java/com/teamcore/site/dao/skill/SkillDAOImpl.java
@@ -32,7 +32,7 @@ public class SkillDAOImpl implements SkillDAO {
             return pst;
         }, keyHolder);
 
-        skill.setId((Long) keyHolder.getKey());
+        skill.setId((Integer)keyHolder.getKey());
 
         return skill;
     }
