+++ src/main/java/com/teamcore/site/domain/DomainObject.java
@@ -2,7 +2,7 @@ package com.teamcore.site.domain;
 
 public interface DomainObject {
 
-    Long getId();
+    Integer getId();
 
-    void setId(Long id);
+    void setId(Integer id);
 }
